\name{launchstRatstat}
\alias{launchstRatstat}
\title{Launch stRat stat in a R shiny viewer}
\description{
This package allows for users to digitize stratigraphic sections and then them into a discretized format that can be further analyzed.}
\usage{launchstRatstat()}
\details{
This package allows for users to digitize stratigraphic sections and then them into a discretized format that can be further analyzed. This process is done through importing an image into stRat stat and hand-picking points on the stratigraphic section. stRat stat includes grain size profiles, sediemntary structures, and three levels of statigraphic heirarchy (i.e., facies, architectral elements, and architectural element sets). When input points are processed, statistics (e.g., min/mean/max grain size, net:gross, thickness) can be computed for each heirarchy level. Additionally both discrete and continuous measurements can be joined to the data and summarized at each heirarchical level.}
\references{https://github.com/ActiveMargins/stRatstat}
\author{Daniel Coutts (dannycoutts@gmail.com, dscoutts@ucalgary.ca)}
\examples{launchstRatstat()}

