% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratstatfunctions.R
\name{launchstRatstat}
\alias{launchstRatstat}
\title{Launch stRat stat}
\usage{
launchstRatstat()
}
\description{
This function launches stRat stat in a Shiny app window. See Github repository (ActiveMargins/stRatstat) for detailed documentation.
}
